Extend referenced output events for finish EDTs
Extend EDT_PROP_OEVT_VALID property to reuse the user provided
event for finish property management.
Both EDT_PROP_OEVT_VALID and EDT_PROP_FINISH properties are supported.
In such case, if the user provided event's type is a latch event, the runtime
will reuse this event to track the finish property. Otherwise, a new latch
event and a dependence between it and the user-provided event will be created.

Change-Id: I60de6481ea90f188d317100867acd60ea3b5fdd8