Update to edtstats analysis ability.
The original method to use this script (ocr/ocr/scripts/EdtStats)
was to enable DEBUG_LVL_VERB - but for many applications this
produced immense log files which would not be postprocessed in a
reasonable time.  This change uses the new Debug_Mask ability to
only enable the necessary DPRINTF statements required by the
edtstats.py script - there by significantly reducing the debug
log files and making them postprocess-able.

The EdtStats/README was also updated with new info for running
on the Xeon Phi processor.

Change-Id: I766d3cf4e042463799935f3a8cbacfe31de26924